Fixing some compiler warnings (mostly relating to missing casts).

+++ Orbitersdk/Space Shuttle Ultra/LCC/SSULCC.cpp
@@ -61,9 +61,9 @@ void SSULCC::clbkPreStep(double simt, double simdt, double mjd)
 	if(timeToLaunch < 0.0) return; // nothing to do
 
 	//if(timeToLaunch>31.0) sprintf(oapiDebugString(),"T %f",-timeToLaunch);
-	int hours = floor(timeToLaunch/3600);
-	int minutes = floor((timeToLaunch - (hours*3600)) / 60);
-	int seconds = floor(timeToLaunch - (hours*3600) - (minutes*60));
+	int hours = static_cast<int>(floor(timeToLaunch/3600));
+	int minutes = static_cast<int>(floor((timeToLaunch - (hours*3600)) / 60));
+	int seconds = static_cast<int>(floor(timeToLaunch - (hours*3600) - (minutes*60)));
 	sprintf(oapiDebugString(),"T -%02i:%02i:%02i",hours,minutes,seconds);
 
 	if(pFSS) {
