+++ Orbitersdk/libUltra/libUltraTests/src/PEG4TargetingTests.cpp
@@ -141,7 +141,7 @@ TEST(PEG4TargetingTest, DeorbitPerturbations) {
 	//VECTOR3 reqdVel = targeting.GetDeltaV() + INITIAL_VEL;
 
 	const double PROPAGATOR_STEP = 1e-2;
-	const double PROPAGATOR_ITERATIONS = 100;
+	const int PROPAGATOR_ITERATIONS = 100;
 	StateVectorPropagator propagator(PROPAGATOR_STEP, PROPAGATOR_ITERATIONS, 3600.0);
 	OMSBurnPropagator omsProp;
 	omsProp.SetBurnData(0.0, targeting.GetDeltaV(), ACCELERATION);
@@ -201,7 +201,7 @@ TEST(PEG4TargetingTest, OMS2Perturbations) {
 	//VECTOR3 reqdVel = targeting.GetDeltaV() + INITIAL_VEL;
 
 	const double PROPAGATOR_STEP = 1e-2;
-	const double PROPAGATOR_ITERATIONS = 100;
+	const int PROPAGATOR_ITERATIONS = 100;
 	StateVectorPropagator propagator(PROPAGATOR_STEP, PROPAGATOR_ITERATIONS, 3600.0);
 	OMSBurnPropagator omsProp;
 	omsProp.SetBurnData(0.0, targeting.GetDeltaV(), ACCELERATION);
@@ -262,7 +262,7 @@ TEST(PEG4TargetingTest, TransferAngle180) {
 	//VECTOR3 reqdVel = targeting.GetDeltaV() + INITIAL_VEL;
 
 	const double PROPAGATOR_STEP = 1e-2;
-	const double PROPAGATOR_ITERATIONS = 100;
+	const int PROPAGATOR_ITERATIONS = 100;
 	StateVectorPropagator propagator(PROPAGATOR_STEP, PROPAGATOR_ITERATIONS, 3600.0);
 	OMSBurnPropagator omsProp;
 	omsProp.SetBurnData(0.0, targeting.GetDeltaV(), ACCELERATION);
